Update the documentation.
In the documentation table of contents rename Example notebooks to Examples.
Remove the contributing info from the notebook pyjanitor_intro.ipynb.
Remove the bold from pyjanitor in the top level title. Doing this to avoid bold in the documentation table of contents.
Change so it is just one "top-level" title in the notebooks pyjanitor_intro and tidytuesday_20190423.ipynb.
Change the order of the table of contents so the pyjaitor_intro notebook always is the first example.
Add Installation in the table of contents.
Change so the examples come before the API Documentation in the table of contents.