[ENH] Add inflation adjustment function #384 (#485)
* Added dictionary of World Bank country names
and their abbreviations for use in the inflate_currency() method.

* Linted World Bank country dictionary in finance.py

* Add helper methods to check that the country name
and the currency_year and to_year variables are valid
for the inflate_currency() method.

* Add inflate_currency() method to finance.py

* Added module of tests for inflate_currency()
finance method.

* Added extra line for linting

* Replaced tabs with spaces

* Linted a line that was too long in
test_inflate_currency.py

* Added more tests to test_inflate_currency.py
to obtain more coverage of the
inflate_currency() method.

* Fixed an odd auto-linting correction that split a
string in two on the same line.

* Removed check for error that should actually
never happen. (This artificially reduced unit
test coverage because a case cannot be
created with the given constraints that would
replicate this issue.)

* Changed country from USA to Ghana (GHA) for
unit tests to check the case where the
`currency_year` or `to_year` are not available
in the API result for the given country.

* Updated the creation of the dictionary
of API results to remove any key:value
items where the value is None

* Fix issue with API to get actual consumer price
index (CPI) with 2010 = 100, rather than
annual CPI increases (as percentage change YOY).

* Updated the doc string for the finance.inflate_currency()
method.

* Removed @deprecated_alias(colname="column_name")
from finance.inflate_currency() method.