test(io): Parameterize test_read_commandline with engine
This commit modifies the test_read_commandline function in the io module to accept an additional parameter, "engine". This allows the function to be tested with different engines, specifically "pandas" and "polars". The function calls within the test have been updated to pass this new parameter. The change is intended to improve the comprehensiveness of the test coverage.